refactor(zimbra): merge diagnostic dnsrecords tests
