+++ packages/manager/apps/zimbra/src/pages/dashboard/Domains/__test__/ModalsDiagnosticDNSRecord.spec.tsx
@@ -5,6 +5,8 @@ import { vi, describe, expect } from 'vitest';
 import { render } from '@/utils/test.provider';
 import { platformMock, domainMock } from '@/api/_mock_';
 import { DomainType } from '@/api/domain';
+import ModalDiagnosticSRV from '../ModalDiagnosticSRV';
+import ModalDiagnosticMX from '../ModalDiagnosticMX';
 import ModalDiagnosticSPF from '../ModalDiagnosticSPF';
 import domainDiagnosticTranslation from '@/public/translations/domains/diagnostic/Messages_fr_FR.json';
 
@@ -14,22 +16,25 @@ vi.mock('@/hooks', () => {
       platformId: platformMock[0].id,
     })),
     useGenerateUrl: vi.fn(),
-    useDomain: vi.fn().mockImplementation((id: string) => ({
+    useDomain: (id: string) => ({
       data: domainMock.find((domain: DomainType) => id === domain.id),
       isLoading: false,
-    })),
+    }),
   };
 });
 
-vi.mock('react-router-dom', () => ({
-  useNavigate: vi.fn(),
-  MemoryRouter: vi.fn(() => <ModalDiagnosticSPF />),
-  useSearchParams: vi.fn(() => [
-    new URLSearchParams({
-      domainId: domainMock[0].id,
-    }),
-  ]),
-}));
+vi.mock('react-router-dom', async (importOriginal) => {
+  const actual = await importOriginal();
+  return {
+    ...actual,
+    useNavigate: vi.fn(),
+    useSearchParams: vi.fn(() => [
+      new URLSearchParams({
+        domainId: domainMock[0].id,
+      }),
+    ]),
+  };
+});
 
 vi.mock('@ovh-ux/manager-react-shell-client', () => ({
   ShellContext: React.createContext({
@@ -56,8 +61,38 @@ afterEach(() => {
   vi.restoreAllMocks();
 });
 
-describe('Domain diagnostic spf modal', () => {
-  it('should be displayed', () => {
+describe('Domain diagnostic modal SRV', () => {
+  it('should display diagnostic modal', () => {
+    const { getByTestId } = render(<ModalDiagnosticSRV />);
+    const modal = getByTestId('modal');
+    expect(modal).toHaveProperty(
+      'headline',
+      domainDiagnosticTranslation.zimbra_domain_modal_diagnostic_srv_title,
+    );
+    expect(getByTestId('diagnostic-srv-modal-secondary-btn')).toBeEnabled();
+  });
+
+  // TODO:
+  // check render when domain is ovh and whn it is not
+});
+
+describe('Domain diagnostic modal MX', () => {
+  it('should display diagnostic modal', () => {
+    const { getByTestId } = render(<ModalDiagnosticMX />);
+    const modal = getByTestId('modal');
+    expect(modal).toHaveProperty(
+      'headline',
+      domainDiagnosticTranslation.zimbra_domain_modal_diagnostic_mx_title,
+    );
+    expect(getByTestId('diagnostic-mx-modal-secondary-btn')).toBeEnabled();
+  });
+
+  // TODO:
+  // check render when domain is ovh and whn it is not
+});
+
+describe('Domain diagnostic modal SPF', () => {
+  it('should display diagnostic modal', () => {
     const { getByTestId } = render(<ModalDiagnosticSPF />);
     const modal = getByTestId('modal');
     expect(modal).toHaveProperty(
